+++ tinymce_4/static/tinymce_4/settings/full/menu.js
@@ -1,6 +1,6 @@
 var tinyMenu = {
   plugins: {
     title: 'Plugins',
-    items: 'insertzoomimage'
+    items: 'insertzoomimage typograf'
   }
 }