+++ CHANGELOG.rst
@@ -1,11 +1,17 @@
 Changelog
 =========
 
-0.0.24
+0.0.26
 ------
 * Update tinymce4 to 5+
 
-0.0.23
+0.0.25
+------
+* Add 'body_class' in settings default class 'content'
+* Update tinymce4 to 4.9.6
+* Update packages and bower deps
+
+0.0.24
 ------
 * Fix the "tinyMCE" initializing for grappelli inline fields
 * Update requirements and bower deps